Access (#99)
* Updating the event policy to filter events restricted to organization members

* Adding the same but for polls

* Misc (#98)

* PollOption has a status column.
* PollOption default  status is in_review
* PollOption suggested by a user are now moderated by openAI
* Minor UI improvements & Code cleaning

* better comments

* Upgrading to octokit 7.0.0 & aws-partitions 1.794.0

* PollOption full moderation

* Removing some debugging code

* Fixing a logic issue allowing poll_options with status different from :approved to be displayed

* Enforcing pundit / rolify on Poll. Refining on Event

* Adding Welcome Policy

* Upgrading net-ssh to version 7.2.0

* Upgrading to flowbite 1.8.0

* Paving the way to invite and restricted access to resource

* Better polls results presentation

---------

+++ app/controllers/polls_controller.rb
@@ -36,16 +36,21 @@ class PollsController < ApplicationController
 
   def edit
     @poll = Poll.find(params[:id])
+    authorize @poll
   end
 
   def show
-    @poll = Poll.find(params[:id])
-    
+    # @poll = Poll.find(params[:id])
+    @poll = Poll.includes(:poll_options).where(poll_options: { status: PollOption.statuses[:approved] }).find(params[:id])
+    authorize @poll
+
     @table_data = VoteCounterService.count_by_poll_option_and_choice(@poll)
+    @user_votes = VoteCounterService.user_votes_for_poll(current_user, @poll)
   end
 
   def destroy
     @poll = current_user.organization.polls.find(params[:id])
+    authorize @poll
     if @poll.destroy
       redirect_to polls_url, notice: "The poll was successfully deleted."
     else
